Weitere Anpassungen bzgl. des Encodings. (refs #4)

+++ Implementierung/Mandatsrechner/src/main/java/gui/ansicht/diagrammfenster/BundDiagramm.java
@@ -116,6 +116,6 @@ public class BundDiagramm extends JPanel {
 	 * @return Dimension
 	 */
 	public Dimension resize() {
-		return new Dimension(this.getWidth(), (int) (this.getHeight()));
+		return new Dimension(this.getWidth(), (this.getHeight()));
 	}
 }
