+++ Implementierung/Mandatsrechner/src/main/java/gui/ansicht/diagrammfenster/DiffDiagramm.java
@@ -104,7 +104,7 @@ public class DiffDiagramm extends JPanel {
 	 * @return Dimension
 	 */
 	public Dimension resize() {
-		return new Dimension(this.getWidth(), (int) (this.getHeight()));
+		return new Dimension(this.getWidth(), (this.getHeight()));
 	}
 
 	/**
