+++ Implementierung/Mandatsrechner/src/main/java/gui/ansicht/diagrammfenster/WahlkreisDiagramm.java
@@ -135,6 +135,6 @@ public class WahlkreisDiagramm extends JPanel {
 	 * @return Dimension
 	 */
 	public Dimension resize() {
-		return new Dimension(this.getWidth(), (int) (this.getHeight()));
+		return new Dimension(this.getWidth(), (this.getHeight()));
 	}
 }
